Perfect Rendering in case of Scaling by a factor

+++ Engg._Drawing_Visualization_n_Design/Source_Code/include/QT/projectionwindow.h
@@ -54,7 +54,7 @@ private:
   int Zval;
 
   // Calculation factor
-  int calc_factor;
+  float calc_factor;
 
   /** Create Actions */
   void createMenus();
